Move keen.js content to home.js.
